feat：博客缓存及点赞缓存完成

+++ ruoyi-modules/ruoyi-wanfeng/src/main/java/com/ruoyi/wanfeng/vo/BlogVo.java
@@ -13,7 +13,7 @@ public class BlogVo extends BaseEntity {
     private String content;
     private String categoryName;
     private String author;
-    private Long likeCount;
+    private int likeCount;
     private Long commentCount;
     private boolean isLike;
 }
